Added optional flags
Added optional flags (CCTV_CAM and FPS) for the Client app.
This lets the user configure which webcam is going to be used by the current Client
and also the FPS that the VideoCapture object is going to use. This is an useful feature,
because different neural networks expect different inputs regarding the frames per second.